requirements.txt
pytest
appium-python-client
PyYAML